Fix some errors, nits and xml2rfc warning
The latest xml2rfc produces a warning about one of the figures being too
wide for the margins, so the figure has been slightly modified. Also, a
previous commit incorrectly identifies the name resolutionr records as
options. And there should be titles and anchors and such to trigger
figure/table numbering, the titles should be as they normally are for
IETF RFCs, etc. (i.e., various nits)