Clean up and fix the description of major and minor versions.
Try to be a bit clearer about when to change the major version and when
to change the minor version.

Fix an error in the description of the minor version (it spoke of code
that reads the new version still being capable of reading the new
version rather than still being capable of reading the *old* version).