Remind people that readers must handle option lists without opt_endofopt.
Indicate that writers MUST put an opt_endofopt at the end (because we
don't want buggy readers that don't also check for the end of the block
to run past the end of the block) and that readers MUST NOT assume that
there will be an opt_endofopt (because we don't want readers to run past
the end of the block if they're reading a file from a buggy - or
malicious! - writer) and SHOULD treat blocks with option lists lacking
an opt_endofopt as if they had an opt_endofopt (so that they can still
read files written by a buggy writer).