Dealing with the two ways to represent 64-bit integers isn't a TODO.
Plenty of pcapng files exist, with 64-bit time stamps written as the
upper 32 bits, in host byte order, followed by the lower 32 bits, in
host byte order, and with other 64-bit quantities written as a 64-bit
number, in host byte order.  There's nothing TO DO about that, other
than live with it, and have writers write files in that fashion and
readers read files in that fashion; fixing it would involve changing the
major version number, and leaving a whole bunch of software out there
incapable of reading the new-format files unless and until they're
updated.

Just note the problem and drive on.