State a rule that must be true for blocks that can contain options.
It must be possible, for those blocks, to determine how much data in the
Block Body comes before the options; otherwise, you couldn't find the
option.

That then implies that you can calculate the number of bytes of options
you have, if any.

Note that this is possible for *all* the standard block types that
support options, and that it must be made true for any type of Custom
Block that supports options; it doesn't require that the data that comes
before the options has a fixed length, just that you have to be able to
calculate how much data that is (for example, from the Captured Length
field in an Enhanced Packet Block, or by scanning all the records in the
Name Resolution Block).