Add an if_hardware option, for a description of the interface hardware.
Use "Broadcom NetXtreme" as an example for if_hardware; give more
"friendly name" strings as examples for if_description, and give some
examples of hardware descriptions.

The idea is to, for example, let programs record both the "friendly
name" and hardware description on Windows, and draw a distinction
between them.  This needs to be cleaned up in both libpcap (in a future
version that supplies IDB-style interface information) and Wireshark.