Mark most strings as not being zero-terminated.
Consistently use the phrase "zero-terminated" when saying that a string
isn't (in most cases) or is (for strings in some NRB records) terminated
by a zero-value octet/NUL character, rather than "null-terminated" or
"without NULL termination".

For strings for which we didn't explicitly say whether the string was
zero-terminated or not, say it's not zero-terminated.