Note restrictions on the lengths of options.
Some options have a fixed length; if the option length isn't the
specified length, the option is invalid.

Some options have a variable length; some of them might have a minimum
length (greater than 0), and if the option length isn't greater than or
equal to the specified length, the option is invalid.

Give the minimum length for variable-length options that have a minimum.

While we're at it, consistently say lower-case-"v" "variable" for
variable-length options.