Fix example times for the ISB.
Use the correct time values for the specified little-endian values.

While we're at it:

Swap the start and end times, as the end time is before the start time.

Use an ISO 8601-like format, rather than an XX/XX/XXXX format, as that
format means MM/DD/YYYY in the US (and Anglophone Canada?) and
DD/MM/YYYY in much of the rest of the world.

Include the fractions-of-a-second time.

Capitalize "Block" in "Interface Description Block".

Note that, in the ISB, the units of the time stamp are as specified in
the IDB referred to by the ISB.