Change the bit order for the bits in epb_flags.
No software that I know of treats bit 0 as the most-significant bit;
Wireshark, and several other programs, treat it as the least-significant
bit, so changing the spec to match reality is probably more practical
than attempting to change reality to match the spec and, in the process,
rendering existing pcapng files with epb_flags options not readable as
intended.

See GitHub issue #57.