Add DSB type for WireGuard (0x57474b4c)
This permits WireGuard handshake keys to be embedded in a packet capture
file to enable decryption in Wireshark. This makes it easier to share
decipherable capture files for educational or debugging purposes.

Implements the functionality requested by
https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=15571

The identifier 0x57474b4c ("WGKL") was arbitrarily chosen.
CRLF support might not be necessary, but is mentioned just in case such
key log text files are processed on Windows.