Give signedness and field size for more values, do other cleanups.
Give the length of more fields; consistently use "N bits" rather than "M
octets" for numerical fields.  Consistently put it immediately after the
field name, rather than in the field description.

Consistently speak of "octets" rather than "bytes".

Give the signedness of numerical fields and options.

While we're at it:

Note that adding a new block type or option is *NOT* a reason to change
the major or minor version number.

Update the name of Apple's UNIX-for-Macs.

Do other copyediting.