Clean up diagrams.
Show the byte offsets for all fields in blocks that are a a fixed
offset.

Align the bit numbers with the bit slots in the diagram, and show the bit
boundaries in all horizontal lines.  Align field separators within words
so they separate bit slots rather than being *in* the bit slots.

Use slashes rather than vertical bars to indicate variable-length
fields.