IDB: add if_txspeed and if_rxspeed options.
Windows supplies separate "transmit speed" and "receive speed" values,
and a value of -1 (0xFFFFFFFFFFFFFFFF) means "not known".  Add
if_txspeed and if_rxspeed options, to handle the case where they differ
or where one is unknown but the other is known; require that if_speed be
used if the speeds are both known and are the same.

I guess they added that for the benefit of asymmetric links such as
ADSL; in practice, few if any machines have built-in ADSL modems, so
you'll probably rarely if ever see this, but at least code that turns
Windows interface information into an IDB has something it can do if
both values aren't present and aren't the same.