Add additional link-layer type values.
Add LINKTYPE_ values that aren't on the tcpdump.org site's page; most of
them are marked as reserved, as people requested them but didn't supply
any specification.

List the range 11-49 as not to be used.  This is to avoid some values
that were assigned to different DLT_ types on different platforms and to
avoid cases where those values might be assigned to DLT_ type in the
future.  (*If* a value is assigned only to one DLT_ type, we could
consider using that value for a corresponding LINKTYPE_, as long as we
can be assured that nobody else will use that value for another
purpose.)

List the range 52-98 as not to be used.  We started assigning values at
100, to avoid both the former problem and to avoid NetBSD assigning
additional values from that range.

While we're at it, note that LINKTYPE_ARCNET_BSD is for ARCNET packets
with BSD's encapsulation (as opposed to Linux's encapsulation).

It Would Be Nice if those who requested these values would supply a
specification; if those types are dissected by tcpdump, Wireshark, or
any other existing software, that specification must match what that
software dissects.

Fixes GitHub issue #102.