"which" and "that" copyedit.
A "restrictive clause" adds additional information to a noun or noun
phrase, so that it identifies a specific item; a "nonrestrictive clause"
adds additional information to a noun or noun phrase that doesn't
restrict what that noun or noun phrase refer to.

Restrictive clauses aren't separated from the noun or noun phrase by a
comma; nonrestrictive clauses are.

At least in US English, restrictive clauses begin with "that" and
nonrestrictive clauses begin with "which".

Replace "that" in a nonrestrictive clause with "which", and add commas
to separate some nonrestrictive clauses.