Take care of version 1.2.
Indicate that readers mus handle version 1.2 the same way that they
handle version 1.0, because some software (improperly) changed the minor
version number, causing existing versions of libpcap and Wireshark not
to be able to read them (libpcap and Wireshark will be updated to handle
them, but there will be non-updated versions of both around for some
amount of time).

Give examples of the sort of file format changes that *would* require a
major version change or a minor version change, and emphasize why
changing the version isn't recommended unless you *really* need it -
i.e., there's going to be a bunch of code using libpcap to read pcapng
files, such as but *not* limited to tcpdump, and other software reading
pcapng files, such as Wireshark, that won't be able to read your Shiny
New Files.