Add a section discussing the version number fields.
Give a longer discussion of that field, the fact that writers must not
write a version number other than 1.0 and *why* that's the case (hint:
you probably don't want to write out files that neither tcpdump nor any
other libpcap-based program nor Wireshark can read), and the
circumstances in which a future version of the specification *might*
specify a version number other than 1.0 (i.e., unless you're making a
change of the sort described here, do *not* propose a version number
change).

Do some copyediting while we're at it.