pcap: fix the description of the field at the end of the file header.
Indicate that it's a 32-bit value, so that it's clear that all 32 bits
must be fetched and put into host byte order.  Describe the meanings of
various bit fields within those 32 bits, which means "*after* the field
is put into host byte order".  This means, for example, that the 16 bits
of LinkType are *not* always at the same offset from the head of the
file - if the file was written in big-endian order, they're at an offset
of 22 bytes from the beginning of the file, whereas, if the file was
written in little-endian order, they're at an offset of 20 bytes from
the beginning of the file.

This fixes GitHub issue #106.