pcap: indicate that a reader SHOULD treat non-zero reserved values as errors.
This is to avoid having code misinterpret captures where those fields,
or parts of them, have been assigned. and non-zero values change the way
that the capture is interpreted, so a program that blithely ignores
non-zero values may misinterpret the packet.  For example, if one of
those reassigned set of bits indicate that some other data should be
interpreted in a different fashion" if it's non-zero, code that ignores
those bits will misinterpret that data if the bits are non-zero.