pcap: show headers a field at a time, without bit numbers.
This should make it clearer, for example, that regardless of the byte
order of the file, the Major Version field is in the two bytes right
after the Magic Number field, and the Minor Version field is in the two
bytes right after the Major Version field.

Consistently speak of "writing" rather than "saving" files.  Do some
other copy-editing.