Redo the way we describe timestamps and the if_tsoffset option.
Describe time stamps in blocks as a 64-bit quantity, with the upper and
lower 32 bits being two parts of that quantity, rather than as two
separate 32-bit quantities.  This makes it easier to use the description
of the timestamp format to describe the value of the start time and end
time options in the Interface Statistics block.

Describe the if_tsoffset option as defaulting to 0 if it's absent, in
the same way that we describe the if_tsresol option as defaulting to 6
(microseconds) if it's absent.  Remove the bit about "the time zone of
the offset" specified by if_tzone; I'm not sure what that was intended
to mean - why would that offset, which appears to be intended to allow
timestamps to be relative to some base time, *have* a time zone? - and
if_tzone isn't sufficient to specify a time zone anyway, so it's been
deprecated and replaced by if_iana_tzname, which contains an IANA time
zone database timezone name.  Remove the bit about an if_tsoffset_low,
as, if somebody wants that, they can request it with an issue or pull
request; that gets rid of one more TODO.

In the full description of the timestamp format in the description of
the Enhanced Packet Block, mention adding the if_tsoffset option to the
timestamp value in the block, making it clearer that readers should do
that. (The Rust pcap_parser crate, at
https://docs.rs/pcap-parser/latest/pcap_parser/, already does so in its
pcapng parsing code.)

In the description of the Interface Statistics Block, don't speak of
that block as a "packet", and indicate that the Interface ID field
specifies the IDB from which the 'if_tsresol' *and* 'if_tsoffset' fields
to use for the IDB timestamp and the start and end time options is to be
taken.