Fix Chaosnet entry.
Fix the capitalization (it was called "Chaosnet" in Dave Moon's memo),
and use the memo in question as a reference. (Point to the Bitsavers
scanned copy; there's a scanned copy at MIT, but the scanning wasn't
very good, and some parts are so faint as to be unreadable.)

Hopefully this will provoke a rebuild, so that the Editor's Copy
reflects the current state of the document.