Explicitly run linktype2stanza with Perl.
It's failing with the extraordinarily unhelpful error "bash: line 1:
./linktype2stanza: cannot execute: required file not found", which
*might* mean that /usr/bin/perl, from the #! line, doesn't exist, but
who knows what the "required file" is?

Explicitly running it with Perl will at least report a "command not
found" error for Perl if there really *is* no Perl interpreter installed
on the Cirrus CI builder or if PATH doesn't include the directory in
which it's installed.