Give up on building linktypes.xml.
Check it into the repository, and remove the last of the attempts to
have it built by the CI process.

If somebody figures out how to cause the build to automatically build
linktyeps.xml from linktypes.csv, including getting Perl, cpanm, and
Text::CSV installed at the right point on the right container, they're
more than welcome to do so and remove linktypes.xml from the build.  all
the shiny wonderful stuff in martinthomson/i-d-template does *NOT*
appear to be easily extensible to add additional steps to the build
process or to easily install the tools necessary for that process.