Add some additional references.
Add a reference for the Xerox experimental 3MB Ethernet; that's not well
documented by any Xerox document I could find at Bitsavers, but this
paper does, at least, include the type field in the header, unlike the
original Ethernet paper.  (That field might not have been used by early
versions of the Ethernet File Transfer Protocol, but I think later
versions ran atop PUP, which did use it, with a type value of 0x0A00.)

Add references for ARCNET; those aren't sufficient to describe the BSD
and Linux encapsulations, which aren't directly described by the ATA
specs, especially when it comes to fragmentation (Linux, I think, might
supply both individual fragments *and* a reassembled frame to the
capture device!), so we leave those as "Reserved for...", which is code
for "we aren't specifying the format here".

While we're at it, move the Radiotap reference to put it in alphabetical
order by reference tag.