linktypes: consistently use "LinkType(s)" to refer to the defined values.
As per a comment in a review of the -05 draft:

https://datatracker.ietf.org/doc/review-ietf-opsawg-pcaplinktype-05-intdir-lc-bernardos-2024-08-22/

> "Linktypes" and LINKTYPE" are used sometimes with the same meaning (I
> guess).

THe term used in the pcap and pcapng I-Ds for the fields containing the
value is LinkType; use that. (That's similar to the Ethernet type field.
values for which are called EtherTypes in IEEE 802.3.)

We note that the name of a LinkType begins with LINKTYPE_ (which is
because they're used as #defines internally to libpcap, and #defines
conventionally use all caps).