pcapng: remove an inconsistency in the description of SPBs.
The description of an SPB has

> The Simple Packet Block does not contain the Interface ID field.
> Therefore, it MUST be assumed that all the Simple Packet Blocks have
> been captured on the interface previously specified in the first
> Interface Description Block.

as its third paragraph; that paragraph seems to imply that an SPB can
appear in a capture with more than one IDB, with the first IDB
describing the interface on which all SPB packets were received or
transmitted.

Later, however, it says

> A Simple Packet Block cannot be present in a Section that has more
> than one interface because of the impossibility to refer to the
> correct one (it does not contain any Interface ID field).

which contradicts that apparent implication.

There's no *inherent* reason why an SPB can't appear in a
multi-interface capture; it's just that you can use SPBs *only* for
packets on that interface - all other packets *must* use an EPB.

State that explicitly as

> As a Simple Packet Block does not contain an Interface ID field, in a
> Section that has more than one interface, only packets received or
> transmitted on the interface described by the first Interface
> Description Block can be contained in a Simple Packet Block; packets
> received or transmitted on any other interface MUST be contained in an
> Enhanced Packet Block.