linktypes: fix some issues in the recent AD review.
No need for an apostrophe in a name for a decade.

Refer to the table of link-layer header types on the tcpdump.org Web
site as "the link-layer header type list".

More accurately describe the history behind the DLT_ values and the
LINKTYPE_ values, which also means it doesn't refer to "operation[sic]
system[s]".