pcap: clarify the requirement for data validation.
Use the terminology - File Header, Packet Records, and Packet Data -
used when describing the format of the file.

Some programs treat the Packet Data in a Packet Record as opaque data,
e.g. tcpdump when using the -w flag and not dissecting and printing
packet contents, and Wireshark's editcap in some cases, so say that the
requirement to validate the Packet Data only applies to programs that
*do* analzye the Packet Data.

Say that those programs must validate *all* of the Packet Data, trusting
*none* of it to be valid; don't just speak of headers in that data.