pcapng: make the description of LinkType more like the pcap I-D version.
Say that it "indicates" rather than "defines" the link layer type of the
interface.

say that the values are defined in the PCAP-related LinkType list
registry, rather than in the link-layer type I-D, as the registry will
be the authoritative list (it may get values added to it after the
link-layer type I-D is published).