draft-ietf-opsawg-pcaplinktype: fix underscore issues.
At least according to [the CommonMark specification section 6.2
"Emphasis and strong
emphasis"](https://spec.commonmark.org/0.31.2/#emphasis-and-strong-emphasis),
"Emphasis with _ is not allowed inside words", so "foo_bar_" as a
paragraph by itself generates the HTML text "<p>foo_bar_</p>".

(If you want to emphasise part of a word, use *, not _.  This may have
been done so as not to cause annoyances with identifiers in code in
languages that allow underscores in identifiers.)

So, at least according to the CommonMark markdown specification, it
shouldn't be necessary to escap eunderscores in LINKTYPE_ names.

However, apparently some Markdown editors don't handle that correctly,
so we'll still escape underscores.

However:

1) We should escape all the underscores in identifiers, not the first
   one, so make the search-and-replace in linktype2stanza global.

2) We have to escape the underscores in the references in the Markdown
   source for the I-D, so that the XML and document preparation process
   can match them.

3) We don't need to use underscores in reference names; use
   Z-WAVE-SERIAL rather than Z_WAVE_SERIAL as the name for a Z-Wave
   document.

4) While we're at it, remove a leading underscore from an entry in
   linktypes.csv for which the corresponding reference didn't have a
   leading underscore.